Split up ThemeContext into different files
